requirements.txt
keras
pandas
Pillow
tensorflow
tk